add index.js
add code for the game, including initialized empty sequence array that keeps track of random keys pressed and the userPos as in the position of the user in the sequence (to check if his sequence is matching so far) as well as the level of the game and the gameStatus which keeps track if the game is still on or not.  Sounds are linked on clicking, and the wrong click is handled.